demo 1 + 2 - reflection + termination strategies/requirements.txt
azure-ai-projects==1.0.0b9
semantic-kernel[azure]==1.28.1
azure-identity